fix: enforce service role key and apply blue theme
