+++ api-test.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>React Project Deployment Guide</title>
+    <title>API Test Console</title>
     <script src="https://cdn.tailwindcss.com"></script>
     <style>
       /* For a nice scrollbar on webkit browsers */
@@ -35,6 +35,6 @@
 </head>
   <body class="bg-[#0E3CB5] text-white">
     <div id="root"></div>
-    <script type="module" src="/index.tsx"></script>
+    <script type="module" src="/api-test.tsx"></script>
   </body>
 </html>
